initial commit (see commit notes)
Initial commit for this project includes a basic prototype for a splat shader
and a 2D test scene that uses the mouse. The system merely takes a single-bit
surface w/ splat team IDs and two 'team color' uniforms to know what to render
given the specified ID.

This is the most basic concept added and will need to be expanded further to
support any kind of texturing that may be required for the 3D pipeline but
that will come once basic 3D rendering is in place.