[addition] splat mesh class
Adds in the splat mesh class which allows rendering a team splat color
to a mesh given a buffer of surface data.

A means to update the surface / buffer is not yet in place but it can
be tested by changing the fill value of the buffer upon mesh creation.