[addition] texture UV paint preview + made things 'work' with Windows export (see commit notes)
GameMaker's consistency between rendering APIs, that being OpenGL and DirectX, is, frankly,
absolute hot garbage.

I'm designing this for Linux first so Windows is an after-thought and, as such, making it
render correctly on Windows has required some janky 'well-enough' adjustments to some areas
to get things working.

y-axis is still upside-down but rays, rendering, and painting is now working correctly.
If I had more knowledge in this area I'd fix things proper but this'll have to do.